Introduce policy boundary
For the roles: connectors secrets manager, console secrets manager,
and cloud storage manager an assume role policy is required that
references the OIDC provider of the k8s cluster. The problem
is that at the time of running this terraform the k8s cluster does
not yet exist.

AWS provides https://aws.amazon.com/blogs/security/when-and-where-to-use-iam-permissions-boundaries/
for just this purpose. Here the customer can define a permissions
boundary for any roles created by the agent. So the agent may create
roles, but they must include this permissions boundary, and the agent
is not permitted to remove this permissions boundary. This will
enable the agent to create the connectors secrets manager, console
secrets manager, and cloud storage manager roles as part of normal
provisioning after the k8s cluster and OIDC provider exist.

For now I am also leaving the connectors secrets manager, console
secrets manager, and cloud storage manager roles here in this terraform
because they are still required by the public API at this time, but
they will be removed as soon as these changes are fully rolled out.