CIAINFRA-289: Bugfix: Lacking permissions for auto-scaling groups
After a recent change to enable subnet creation to occur outside
of this package it caused an issue where the `ec2:RunInstances` permission
was no longer granted on the subnet resources.

With this change the user will be required to pass in either subnet
cidrs, in which case we will create the subnets, or a list of subnet ids
if created externally.

Either way the resulting subnets are identified and `ec2:RunInstances`
will be granted on the subnets.