Minor changes to default behavior
I'd like the default behavior of the code to more closely aligned
with what our customers commonly need.

1. Remove `public_cluster` variable. Customer-Managed VPC clusters are never public.
2. Default the list of `public_subnet_cidrs` to be empty. All subnets are typically
created external to this code. The primary use case for `public_subnet_cidrs` is our
own internal testing.
3. Add new variable `vpc_cidr_block` which is only used when `vpc_id` is not provided.
Since it is possible to provide a list of `private_subnet_cidrs` it should also be
possible to set the CIDR block for the VPC if it is being created.